Fixed CORS error on requests to permalink service
