+++ modules/server/src/main/scala/es/weso/server/LinksService.scala
@@ -1,45 +1,17 @@
 package es.weso.server
+
 import cats.effect._
-import es.weso.server.QueryParams.UrlParam
 import org.http4s.Uri.uri
 import org.http4s._
 import org.http4s.dsl.Http4sDsl
 import org.http4s.headers.Location
-import scalaj.http.Http
-import play.api.libs.json._
 
 class LinksService[F[_]](blocker: Blocker)(implicit F: Effect[F], cs: ContextShift[F])
   extends Http4sDsl[F] {
 
-  val urlShortenerEndpoint = "https://api-ssl.bitly.com/v4/shorten"
-  val urlShortenerDomain = "bit.ly"
-
   case class RequestData(domain: String, url: String)
 
   val routes: HttpRoutes[F] = HttpRoutes.of[F] {
-
-    // Query URL shortener API
-    case GET -> Root / "links" / "shorten" :?
-      UrlParam(url) =>
-      // Request the shortened URL
-      try {
-        val res = Http(urlShortenerEndpoint).postData(s"""{ "long_url": "$url" }""")
-          .header("Content-Type", "application/json")
-          .header("Accept", "application/json")
-          .header("Authorization", "8e4c5fc54b280e5bc969fc0ef7a0f759317b7c64")
-          .asString
-
-        val jsonRes = Json.parse(res.body)
-
-        // Return the short URL if possible
-        if (res.code >= 200 && res.code < 300 && (jsonRes \ "link").isDefined)
-          Ok(s"""{ "og_link": "$url", "short_link": "${(jsonRes \ "link").as[String]}" }""")
-        else
-          InternalServerError (url)
-      }
-      catch {
-        case e: ArithmeticException => InternalServerError (url)
-      }
 
     // data Info
     case GET -> Root / "links" / "i1" =>
