+++ modules/server/src/main/scala/es/weso/server/Server.scala
@@ -1,22 +1,15 @@
 package es.weso.server
-import org.http4s._
-import org.http4s.implicits._
-import org.http4s.server._
-import org.http4s.server.blaze.BlazeServerBuilder
-import org.http4s.server.middleware.{CORS, HSTS, Logger}
-import es.weso.server.utils.Http4sUtils._
-import org.log4s.getLogger
 import cats.effect._
 import cats.implicits._
-import es.weso.quickstart.{HelloWorld, TestRoutes}
 import fs2.Stream
-import javax.net.ssl.SSLContext
+import org.http4s._
 import org.http4s.client.Client
 import org.http4s.client.blaze.BlazeClientBuilder
-import org.http4s.dsl.Http4sDsl
+import org.http4s.implicits._
+import org.http4s.server.blaze.BlazeServerBuilder
+import org.http4s.server.middleware.{CORS, Logger}
 
 import scala.concurrent.ExecutionContext.global
-import scala.util.Properties.envOrNone
 import scala.concurrent.duration._
 
 object Server {
@@ -31,7 +24,6 @@ object Server {
   } */
 
   def routesService[F[_]: ConcurrentEffect](blocker: Blocker, client: Client[F])(implicit T: Timer[F], C: ContextShift[F]): HttpRoutes[F] =
-//    HelloService[F](blocker).routes <+>
     CORS (
       SchemaService[F](blocker,client).routes <+>
       APIService[F](blocker, client).routes <+>
@@ -39,7 +31,9 @@ object Server {
       ShExService[F](blocker,client).routes <+>
       ShapeMapService[F](blocker,client).routes <+>
       WikidataService[F](blocker, client).routes <+>
-      EndpointService[F](blocker,client).routes
+      EndpointService[F](blocker,client).routes <+>
+      PermalinkService[F](blocker,client).routes
+
     ) <+>
     WebService[F](blocker).routes <+>
     DataWebService[F](blocker, client).routes <+>
